chore(docker_api): separate container related code
