+++ src/docker_api/mod.rs
@@ -1,1 +1,2 @@
 pub mod client;
+pub mod container;
